fix(playerbot): Move PetMetrics to namespace scope and remove duplicate
Similar to EquipmentMetrics, PetMetrics was defined as nested struct in both
interface and implementation causing covariant return type error. Moved to
Playerbot namespace scope with atomic members and removed duplicate definition
from BattlePetManager.h.
